Feature/notification

+++ apps/backend/src/like/like.service.ts
@@ -1,64 +1,62 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { NotificationGateway } from '../notification/notification.gateway';
+import { NotificationType } from '@prisma/client';
 
 @Injectable()
 export class LikeService {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    private notificationGateway: NotificationGateway,
+  ) {}
 
   async likePost(postId: string, userId: string) {
-    // si l'user n'a pas déjà liké le post
-    if(await this.prisma.like.findFirst({
-      where: {
-        postId: postId,
-        userId: userId,
-      },
-    })) {
+    if (await this.prisma.like.findFirst({ where: { postId, userId } })) {
       return { message: 'Post already liked' };
     }
-    await this.prisma.like.create({
+    const like = await this.prisma.like.create({
       data: {
         post: { connect: { id: postId } },
         user: { connect: { id: userId } },
       },
     });
-    return { message: 'Post liked', postId: postId };
+
+    const post = await this.prisma.post.findUnique({ where: { id: postId }, include: { user: true } });
+    if (post.userId !== userId) {
+      const message = `${post.user.username} a aimé votre post`;
+      await this.notificationGateway.sendNotificationToUser(post.userId, NotificationType.LIKE, message, userId);
+    }
+
+    return { message: 'Post liked', postId };
   }
 
   async unlikePost(postId: string, userId: string) {
-    await this.prisma.like.deleteMany({
-      where: {
-        postId: postId,
-        userId: userId,
-      },
-    });
-    return { message: 'Post unliked', postId: postId };
+    await this.prisma.like.deleteMany({ where: { postId, userId } });
+    return { message: 'Post unliked', postId };
   }
 
   async likeComment(commentId: string, userId: string) {
-    if(await this.prisma.like.findFirst({
-      where: {
-        commentId: commentId,
-        userId: userId,
-      },
-    })) {
+    if (await this.prisma.like.findFirst({ where: { commentId, userId } })) {
       return { message: 'Comment already liked' };
     }
-    await this.prisma.like.create({
+    const like = await this.prisma.like.create({
       data: {
         comment: { connect: { id: commentId } },
         user: { connect: { id: userId } },
       },
     });
-    return { message: 'Comment liked', commentId: commentId };
+
+    const comment = await this.prisma.comment.findUnique({ where: { id: commentId }, include: { user: true } });
+    if (comment.userId !== userId) {
+      const message = `${comment.user.username} a aimé votre commentaire`;
+      await this.notificationGateway.sendNotificationToUser(comment.userId, NotificationType.LIKE, message, userId);
+    }
+
+    return { message: 'Comment liked', commentId };
   }
 
   async unlikeComment(commentId: string, userId: string) {
-    await this.prisma.like.deleteMany({
-      where: {
-        commentId: commentId,
-        userId: userId,
-      },
-    });
-    return { message: 'Comment unliked', commentId: commentId };
+    await this.prisma.like.deleteMany({ where: { commentId, userId } });
+    return { message: 'Comment unliked', commentId };
   }
 }
